+++ packages/integration-sdk-testing/src/__tests__/jest.test.ts
@@ -18,7 +18,7 @@ import {
   registerMatchers,
   toImplementSpec,
 } from '../jest';
-import { v4 as uuid } from 'uuid';
+import { randomUUID as uuid } from 'crypto';
 import { toMatchStepMetadata } from '..';
 import { getMockIntegrationStep } from '@jupiterone/integration-sdk-private-test-utils';